% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-lab-meta.R
\name{assign_mek_loc}
\alias{assign_mek_loc}
\title{assign a location on the robot table for a destination or source plate}
\usage{
assign_mek_loc(plate_names, table, dest_or_source, identifier)
}
\arguments{
\item{plate_names}{names of plates to be used on the robot}

\item{table}{table of samples}

\item{dest_or_source}{is this plate a source or a destination}

\item{identifier}{digest_id or ligation_id}
}
\description{
assign a location on the robot table for a destination or source plate
}
\examples{
source <- assign_mek_loc(dig_plates, source, "source", "digest_id")
}
\author{
Michelle Stuart
}
