% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-gen.R
\name{change_db_gen_id}
\alias{change_db_gen_id}
\title{change the gen_id field in the database for all samples in a table}
\usage{
change_db_gen_id(table_of_candidates)
}
\arguments{
\item{x}{= table_of_candidates}
}
\description{
change the gen_id field in the database for all samples in a table
}
\examples{
change_db_gen_id(table_of_candidates)
}
\author{
Michelle Stuart
}
