% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-sample-meta.R
\name{change_rows}
\alias{change_rows}
\title{once rows have been changed remove them from table and add them back in}
\usage{
change_rows(table, change, identifier)
}
\arguments{
\item{x}{= whole table}

\item{y}{= the changed rows}

\item{z}{= identifying co}
}
\description{
once rows have been changed remove them from table and add them back in
}
\examples{
deer <- change_rows(deer, change)
}
\author{
Michelle Stuart
}
