% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-gen.R
\name{check_id_match}
\alias{check_id_match}
\title{compare individuals that appear to be genetic mark recaptures}
\usage{
check_id_match(row_number, table_of_matches)
}
\arguments{
\item{x}{= table_name}
}
\description{
compare individuals that appear to be genetic mark recaptures
}
\examples{
temp <- check_id_match(row_number, table_of_matches)
}
\author{
Michelle Stuart
}
