% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-gen.R
\name{create_genid}
\alias{create_genid}
\title{add a gen_id to a sample}
\usage{
create_genid(table_of_candidates)
}
\arguments{
\item{x}{= table_name}
}
\description{
add a gen_id to a sample
}
\examples{
updated_table <- create_genid(table_of_candidates)
}
\author{
Michelle Stuart
}
