% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-sample-meta.R
\name{daterange_dive}
\alias{daterange_dive}
\title{for a given date range, list all of the dives}
\usage{
daterange_dive(begin_date, end_date)
}
\arguments{
\item{begin_date}{the beginning of the range}

\item{end_date}{the end of the range}
}
\description{
for a given date range, list all of the dives
}
\examples{
dat <- daterange_dive("2016-01-01", "2017-12-30")
}
\author{
Michelle Stuart
}
