% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-sample-meta.R
\name{fish_anem_dive}
\alias{fish_anem_dive}
\title{fish_anem_dive grabs this data for fish from the db}
\usage{
fish_anem_dive()
}
\value{
a table of all fish, associated anemones, and dive info
}
\description{
fish_anem_dive grabs this data for fish from the db
}
\examples{
fish_meta <- fish_anem_dive()
}
