% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-sample-meta.R
\name{get_anem}
\alias{get_anem}
\title{Get anemone data}
\usage{
get_anem()
}
\value{
the anemone db table
}
\description{
Get anemone data
}
\examples{
anem_17 <- get_anem() \%>\% filter(anem_id > 2000)
}
