% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-lab-meta.R
\name{get_dig}
\alias{get_dig}
\title{Get Dig}
\usage{
get_dig()
}
\value{
a table of digests with meta data
}
\description{
Get Dig
}
\examples{
test <- get_dig() \%>\% select(digest_id, extraction_id)

}
