% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-lab-meta.R
\name{get_lig}
\alias{get_lig}
\title{Get Lig}
\usage{
get_lig()
}
\arguments{
\item{ligation_ids}{a table that contains ligation_ids}
}
\value{
a table of ligation ids with meta data
}
\description{
Get Lig
}
\examples{
test <- get_lig(table$ligation_id)
}
