% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-lab-meta.R
\name{heatmap}
\alias{heatmap}
\title{heatmap - plot a plate map with color}
\usage{
heatmap(plate_as_long_table, id)
}
\arguments{
\item{plate_as_long_table}{a table of samples and plate locations}

\item{id}{sample identifier}
}
\description{
heatmap - plot a plate map with color
}
\examples{
sample_map <- heatmap(plate_as_long_table)
}
\author{
Michelle Stuart
}
