% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-gen.R, R/funs-lab-meta.R
\name{lig_from_samp}
\alias{lig_from_samp}
\title{views all of the fish recaptured at a given site}
\usage{
lig_from_samp(sample_ids)

lig_from_samp(sample_ids)
}
\arguments{
\item{x}{= list of sample_ids}

\item{x}{= list of sample_ids}
}
\description{
views all of the fish recaptured at a given site

views all of the fish recaptured at a given site
}
\examples{
fish <- lig_from_samp(c("APCL13_516", "APCL13_517"))
fish <- lig_from_samp(c("APCL13_516", "APCL13_517"))
}
\author{
Michelle Stuart

Michelle Stuart
}
