% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-lab-meta.R
\name{lig_ng}
\alias{lig_ng}
\title{lig_ng figure out how many ng to use in making pools for ligations}
\usage{
lig_ng(dig, regeno)
}
\arguments{
\item{dig}{a table of digests that need to be ligated}

\item{regeno}{a table of samples to be regenotyped}
}
\value{
a table of ligations with volumes of sample and water to combine
}
\description{
lig_ng figure out how many ng to use in making pools for ligations
}
\examples{
ligs <- lig_ng(dig)
}
\author{
Michelle Stuart
}
