% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-lab-meta.R
\name{make_plate}
\alias{make_plate}
\title{make a plate from a list of sample_ids, extraction_ids, etc.}
\usage{
make_plate(list_of_ids)
}
\arguments{
\item{list_of_ids}{a list of ids}
}
\value{
a table of plate locations for samples
}
\description{
make a plate from a list of sample_ids, extraction_ids, etc.
}
\examples{
test <- make_plate(lig_ids)
}
\author{
Michelle Stuart
}
