% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-lab-meta.R
\name{make_plate_with_negs}
\alias{make_plate_with_negs}
\title{Title}
\usage{
make_plate_with_negs(list_of_ids, id_type)
}
\arguments{
\item{list_of_ids}{a list of ids}

\item{id_type}{character string of id column name}
}
\value{
a table of ids and well locations
}
\description{
Title
}
\examples{
plate <- make_plate_with_negs(samples, "extraction_id")
}
