% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-lab-meta.R
\name{make_platemap}
\alias{make_platemap}
\title{Make Platemap}
\usage{
make_platemap(tble)
}
\arguments{
\item{tble}{a table of samples with well locations}
}
\value{
a platemap
}
\description{
Make Platemap
}
\examples{
platmap <- make_platemap(plate)
}
