% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-lab-meta.R
\name{plate_from_db}
\alias{plate_from_db}
\title{plate_from_db recreates the platemap based on well locations in the db}
\usage{
plate_from_db(table_name, ...)
}
\arguments{
\item{table_name}{a table imported from the database that includes id and well columns}

\item{...}{id to be displayed on the plate map}
}
\value{
a platemap representing where samples went in a plate
}
\description{
plate_from_db recreates the platemap based on well locations in the db
}
\examples{
platemap <- plate_from_db(extractions, sample_id)
}
