% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-sample-meta.R
\name{read_db}
\alias{read_db}
\title{read_db is a helper function to connect to db with read only intent}
\usage{
read_db(x)
}
\arguments{
\item{x}{= which db?}
}
\description{
read_db is a helper function to connect to db with read only intent
}
\examples{
db <- read_Db("Leyte")
}
\author{
Michelle Stuart
}
