% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-gen.R
\name{read_genepop}
\alias{read_genepop}
\title{read a genepop generated for ONE population}
\usage{
read_genepop(filename)
}
\arguments{
\item{x}{= filename}
}
\description{
read a genepop generated for ONE population
}
\examples{
genedf <- read_genepop("data/seq17_03_58loci.gen")
}
\author{
Michelle Stuart
}
