% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-lab-meta.R
\name{remove_rows}
\alias{remove_rows}
\title{Remove rows}
\usage{
remove_rows(table_name, how_many_wells)
}
\arguments{
\item{table_name}{table of samples}

\item{how_many_wells}{how many wells in the plate}
}
\value{
a sliced table
}
\description{
Remove rows
}
\examples{
samp <- remove_rows(samp, 48)

}
