% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-lab-meta.R
\name{samp_from_lig}
\alias{samp_from_lig}
\title{find sample id from ligation id}
\usage{
samp_from_lig(table_name)
}
\arguments{
\item{table_name}{table containing ligation ids}
}
\description{
find sample id from ligation id
}
\examples{
c5 <- samp_from_lig(genedf)
}
\author{
Michelle Stuart
}
