% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-sample-meta.R
\name{sample_latlon}
\alias{sample_latlon}
\title{sample_latlon}
\usage{
sample_latlon(sample_ids)
}
\arguments{
\item{sample_ids}{}
}
\value{
a table of sample_ids with lat lons
}
\description{
sample_latlon
}
\examples{
where <- sample_latlon(c("APCL12_043", "APCL17_267"))

}
