% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-sample-meta.R
\name{site_recap}
\alias{site_recap}
\title{site_recap finds recaptured fish by site}
\usage{
site_recap(site_name)
}
\arguments{
\item{site_name}{name of site of interest}
}
\value{
a table of fish that have been recaptured at a given site
}
\description{
site_recap finds recaptured fish by site
}
\examples{
recaps <- site_recap("Haina")
}
