% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-lab-meta.R
\name{work_history}
\alias{work_history}
\title{Title}
\usage{
work_history(table, id_column)
}
\arguments{
\item{table}{table_where_ids_are}

\item{id_column}{column_of_ids - must be sample_id, extraction_id, digest_id, or ligation_id}
}
\value{
a table of lab metadata
}
\description{
Title
}
\examples{

}
