% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-sample-meta.R
\name{write_db}
\alias{write_db}
\title{write_db establishes a connection with the mysql db with intent to change it}
\usage{
write_db(db_name)
}
\arguments{
\item{db_name}{}
}
\value{
database connection
}
\description{
write_db establishes a connection with the mysql db with intent to change it
}
\examples{
leyte <- write_db("Leyte")
}
