% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-sample-meta.R
\name{anem_dive}
\alias{anem_dive}
\title{for a given anemone , returns site, date, divetype}
\usage{
anem_dive(anems_ids)
}
\arguments{
\item{anem_ids}{anem(s) for which to get data}
}
\description{
for a given anemone , returns site, date, divetype
}
\examples{
info <- anem_dive(2183)
}
\author{
Michelle Stuart
}
